Fix wrapping with arrows & sporadic seg fault
Used a CompoundDrawable in a TextView instead of the ImageView &
TextView combo to allow both wrapping and resizing to the full item
width.

The seg fault was caused in the C code interaction with the camera. It
appeared to be an out of memory issue (some allocation seemed to receive
a null pointer eventually). Fix was to slow down the ImageReader rate by
an arbitrary factor (5 was used).